refactor: CookieManager 유틸 클래스로 변경 - CookieManager client 패키지로 이동
