+++ src/main/java/roomescape/auth/interceptor/AdminAuthInterceptor.java
@@ -4,12 +4,11 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
-import roomescape.auth.CookieManager;
 import roomescape.auth.controller.AuthController;
 import roomescape.auth.service.AuthService;
+import roomescape.client.CookieManager;
 import roomescape.exception.ExceptionMessage;
 import roomescape.exception.ForbiddenException;
-import roomescape.exception.UnAuthorizedException;
 import roomescape.member.domain.Member;
 
 @Component
@@ -23,8 +22,7 @@ public class AdminAuthInterceptor implements HandlerInterceptor {
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
-        CookieManager cookieManager = new CookieManager(request.getCookies());
-        String accessToken = cookieManager.getValue(AuthController.AUTH_TOKEN_COOKIE);
+        String accessToken = CookieManager.getValue(request.getCookies(), AuthController.AUTH_TOKEN_COOKIE);
 
         Member member = authService.getLoginMember(accessToken);
         if (member.isAdmin()) {
