+++ src/main/java/roomescape/auth/resolver/LoginMemberArgumentResolver.java
@@ -7,10 +7,10 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import roomescape.auth.CookieManager;
 import roomescape.auth.controller.AuthController;
 import roomescape.auth.controller.LoginMember;
 import roomescape.auth.service.AuthService;
+import roomescape.client.CookieManager;
 import roomescape.member.domain.Member;
 
 @Component
@@ -31,8 +31,7 @@ public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolve
     public LoginMember resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
         HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
 
-        CookieManager cookieManager = new CookieManager(request.getCookies());
-        String accessToken = cookieManager.getValue(AuthController.AUTH_TOKEN_COOKIE);
+        String accessToken = CookieManager.getValue(request.getCookies(), AuthController.AUTH_TOKEN_COOKIE);
         Member member = authService.getLoginMember(accessToken);
         return new LoginMember(member.getId(), member.getName(), member.getEmail(), member.getRole());
     }
